+++ graph_attention_student/experiments/vgd_single_megan_tadf_ktadf.py
@@ -2,6 +2,7 @@ import os
 import pathlib
 import typing as t
 
+import numpy as np
 import tensorflow as tf
 import tensorflow.keras as ks
 # from pycomex.util import Skippable
@@ -10,9 +11,9 @@ from pycomex.functional.experiment import Experiment
 from pycomex.utils import folder_path, file_namespace
 
 # == DATASET PARAMETERS ==
-VISUAL_GRAPH_DATASET_PATH: str = os.path.expanduser('/media/ssd/.visual_graph_datasets/datasets/aqsoldb')
-USE_DATASET_SPLIT: t.Optional[int] = 0
-TRAIN_RATIO: float = 0.8
+VISUAL_GRAPH_DATASET_PATH: str = os.path.expanduser('/media/ssd/.visual_graph_datasets/datasets/tadf')
+USE_DATASET_SPLIT: t.Optional[int] = None
+TRAIN_RATIO: float = 0.99
 NUM_EXAMPLES: int = 100
 NUM_TARGETS: int = 1
 
@@ -24,29 +25,31 @@ USE_EDGE_LENGTHS: bool = False
 USE_EDGE_ATTRIBUTES: bool = True
 
 # == MODEL PARAMETERS ==
-UNITS = [48, 48, 48]
+UNITS = [64, 64, 64, 64]
 DROPOUT_RATE = 0.1
-FINAL_UNITS = [48, 32, 16, 1]
-REGRESSION_REFERENCE = [[-3.2]]
+FINAL_UNITS = [64, 32, 16, 1]
+REGRESSION_REFERENCE = [[-22.]]
 REGRESSION_WEIGHTS = [[1.0, 1.0]]
 IMPORTANCE_CHANNELS: int = 2
 IMPORTANCE_FACTOR = 2.0
-IMPORTANCE_MULTIPLIER = 0.6
+IMPORTANCE_MULTIPLIER = 0.8
 FIDELITY_FACTOR = 0.2
 FIDELITY_FUNCS = [
     lambda org, mod: tf.nn.relu(mod - org),
     lambda org, mod: tf.nn.relu(org - mod),
 ]
-SPARSITY_FACTOR = 0.3
+SPARSITY_FACTOR = 0.1
 CONCAT_HEADS = False
 
 # == TRAINING PARAMETERS ==
+DEVICE = 'cpu:0'
 BATCH_SIZE = 32
-EPOCHS = 75
+EPOCHS = 25
 REPETITIONS = 1
 OPTIMIZER_CB = lambda: ks.optimizers.experimental.AdamW(learning_rate=0.001)
 
 # == EXPERIMENT PARAMETERS ==
+LOG_STEP = 10_000
 __DEBUG__ = True
 __TESTING__ = False
 
@@ -56,5 +59,14 @@ experiment = Experiment.extend(
     namespace=file_namespace(__file__),
     glob=globals()
 )
+
+
+@experiment.hook('get_graph_labels')
+def get_graph_labels(e, metadata, graph):
+    value = float(metadata['target'][2])
+    if value < 1e-30:
+        raise ValueError()
+    return np.log10(np.expand_dims(np.array(value), axis=0))
+
 
 experiment.run_if_main()